+++ code/monster.h
@@ -8,15 +8,15 @@ class Monster
 public:
 
     Monster(std::string name_, int level);
-    int GetStat(const std::string &stat);
-    std::map<std::string, int> GetStats();
+    float GetStat(const std::string &stat);
+    std::map<std::string, float> GetStats();
 
     std::string name;
     int level;
     std::string element;
-    std::map<std::string, int> base_stats;
+    std::map<std::string, float> base_stats;
     int level_up{};
     int xp{};
-    int health{};
-    int energy{};
+    float health{};
+    float energy{};
 };
